refector : float 형변환을 수익률 계산 시에만 하도록 변경

+++ src/main/java/autolotto/model/ProfitCalculator.java
@@ -8,6 +8,6 @@ public class ProfitCalculator {
         for (int i = 3; i < 7; i++) {
             profit = ProfitMatcher.getMatchProfit(i, lottosComparer.getMatchingCount(i));
         }
-        return profit / money;
+        return profit / (float)money;
     }
 }
